v3

+++ main.py
@@ -83,7 +83,6 @@ async def get_prints_from_token(token : str, invalidate_cache : bool = False) ->
     global CACHE
 
     if token in CACHE and not invalidate_cache and (time.time() < (CACHE[token]['time'] + 3600)):
-        print('CACHE HIT!')
         return CACHE[token]['data']
 
     async with aiohttp.ClientSession() as session:
@@ -106,6 +105,9 @@ async def get_prints_from_token(token : str, invalidate_cache : bool = False) ->
 
     return CACHE[token]['data']
 
+async def uid_in_prints_from_token(token : str, uid : str) -> bool:
+    items = await get_prints_from_token(token)
+    return any(x['uid'] == uid for x in items)
 
 # ---
 # Utils
@@ -152,11 +154,13 @@ async def uid_download_embed(uid : str, color : int = 0x00FF00) -> discord.Embed
 def extract_uid(url : str) -> str:
     uid = None
 
-    if (url.startswith('https://www.thingiverse.com/thing:')):
+    if url.startswith('prusa-printables:') or url.startswith('thingiverse:') or url.startswith('myminifactory:'):
+        uid = url
+    elif url.startswith('https://www.thingiverse.com/thing:'):
         uid = f"thingiverse:{url.split(':')[-1]}"
-    elif (url.startswith("https://www.myminifactory.com/object/")):
+    elif url.startswith("https://www.myminifactory.com/object/"):
         uid = f"myminifactory:{url.split('-')[-1]}"
-    elif (url.startswith("https://www.printables.com/model")):
+    elif url.startswith("https://www.printables.com/model"):
         uid = f"prusa-printables:{url.split('/')[-1].split('-')[0]}"
 
     return uid
@@ -166,41 +170,23 @@ def extract_uid(url : str) -> str:
 # ---
 
 class InteractButton(discord.ui.View):
-    def __init__(self, uid : str, user_id: int):
-        super().__init__(timeout=86400)
-        self.uid = uid
-        self.user_id = user_id
-        self.completed = False
-
-    @discord.ui.button(label='Complete', style=discord.ButtonStyle.danger)
-    async def confirm(self, interaction: discord.Interaction, button: discord.Button):
-        if (interaction.user.id != self.user_id):
-            await interaction.response.send_message('This button belongs to someone else.', ephemeral=True)
-            return
-
-        if (self.completed):
-            await interaction.response.send_message('This print has already been completed.', ephemeral=True)
-            return
-
-        await print_complete(interaction, self.uid)
-        self.completed = True
-
-    @discord.ui.button(label='Download', style=discord.ButtonStyle.primary)
-    async def list_downloads(self, interaction : discord.Interaction, button : discord.Button):
-        await interaction.response.defer(ephemeral=True)
-        embed = await uid_download_embed(self.uid)
-        await interaction.followup.send(embed=embed, ephemeral=True)
-
-class OnlyDownloadButton(discord.ui.View):
     def __init__(self, uid : str):
         super().__init__(timeout=86400)
         self.uid = uid
 
+    @discord.ui.button(label='Complete', style=discord.ButtonStyle.danger)
+    async def confirm(self, interaction: discord.Interaction, button: discord.Button):
+        await print_complete(interaction, self.uid)
+
     @discord.ui.button(label='Download', style=discord.ButtonStyle.primary)
     async def list_downloads(self, interaction : discord.Interaction, button : discord.Button):
         await interaction.response.defer(ephemeral=True)
         embed = await uid_download_embed(self.uid)
         await interaction.followup.send(embed=embed, ephemeral=True)
+
+    @discord.ui.button(label='Add to queue', style=discord.ButtonStyle.secondary)
+    async def add_to_queue(self, interaction : discord.Interaction, button : discord.Button):
+        await print_add(interaction, self.uid)
 
 # ---
 # Commands
@@ -214,10 +200,9 @@ async def posts_autocomplete(interaction: discord.Interaction, current: str) -> 
             discord.app_commands.Choice(name=x['name'], value=x['uid'])
             for x in items if current.lower() in x['name'].lower()
         ]
-        print(result)
     except Exception as e:
         result = []
-        print(str(e))
+        logger.error(str(e))
 
     return result
 
@@ -229,6 +214,10 @@ async def print_complete_command(interaction: discord.Interaction, print_name : 
 async def print_complete(interaction: discord.Interaction, uid : str, show_in_channel : bool = False):
     await interaction.response.defer(ephemeral=not show_in_channel)
     token = await get_channel_mapping(str(interaction.user.id), interaction.user.name)
+    
+    if not await uid_in_prints_from_token(token, uid):
+        await interaction.followup.send('Print not found in queue', ephemeral=True)
+        return
 
     async with aiohttp.ClientSession() as session:
         async with session.delete(base_url + f'/Saved/{token}/remove', json={
@@ -243,6 +232,9 @@ async def print_complete(interaction: discord.Interaction, uid : str, show_in_ch
     await interaction.followup.send('Print completed, removed from queue', embed=embed, ephemeral=not show_in_channel)
 
 @bot.print_group.command(name='add', description='Add a 3d print URL to the queue')
+async def print_add_command(interaction: discord.Interaction, url : str, show_in_channel : bool = False):
+    await print_add(interaction, url, show_in_channel)
+
 async def print_add(interaction: discord.Interaction, url : str, show_in_channel : bool = False):
     url = url.strip()
     await interaction.response.defer(ephemeral=not show_in_channel)
@@ -262,40 +254,43 @@ async def print_add(interaction: discord.Interaction, url : str, show_in_channel
         }) as response:
             success = response.status == 200
             fail = await response.text()
-            print(fail, response.status)
+            if not success:
+                logger.error(fail + ' ' + str(response.status))
 
     asyncio.create_task(get_prints_from_token(token, True))
-    view = InteractButton(uid, interaction.user.id)
+    view = InteractButton(uid)
 
     if success:
         embed = await uid_embed(uid)
         if not show_in_channel:
             embed.set_footer(text=f'API Code: {token}')
 
-        await interaction.followup.send('Done!', embed=embed, view=view, ephemeral=not show_in_channel)
+        await interaction.followup.send('Added print to queue.', embed=embed, view=view, ephemeral=not show_in_channel)
     else:
         await interaction.followup.send(f'Failed! {fail}', ephemeral=True)
 
 @bot.print_group.command(name='list', description='List current 3d print files in queue')
 @discord.app_commands.autocomplete(print_name=posts_autocomplete)
-async def print_list(interaction: discord.Interaction, print_name : str = None, show_in_channel : bool = False):
+async def print_list_command(interaction: discord.Interaction, print_name : str = None, show_in_channel : bool = False):
+    await print_list(interaction, print_name, show_in_channel)
+
+async def print_list(interaction: discord.Interaction, uid : str = None, show_in_channel : bool = False):
     await interaction.response.defer(ephemeral=not show_in_channel)
-    
     token = await get_channel_mapping(str(interaction.user.id), interaction.user.name)
     items = await get_prints_from_token(token)
 
-    if (print_name != None):
-        embed = await uid_embed(print_name)
+    if (uid != None):
+        embed = await uid_embed(uid)
 
         if not show_in_channel:
             embed.set_footer(text=f'API Code: {token}')
 
-        view = InteractButton(print_name, interaction.user.id)
+        view = InteractButton(uid)
         await interaction.followup.send(embed=embed, view=view, ephemeral=not show_in_channel)
         return
 
     embed = discord.Embed(title='Queued Items', color=0xFFFF00, description='\n'.join(
-        f"- [{x['name']}]({x['url']})" for x in items
+        f"- {x['name']}" for x in items
     ))
 
     if not show_in_channel:
@@ -313,13 +308,18 @@ async def print_help(interaction : discord.Interaction):
         f'`/print {x.name}` - {x.description}' for x in bot.print_group.commands
     ), inline=False)
 
+    embed.add_field(name="How to use", value="This bot keeps track of a list of prints for you. You can use this list however you like, using it as a printing queue is just a suggestion.\n- You can add prints to this list with `/print add <url>`\n- You can view the list with `/print list`\n- You can view a specific print with `/print list print_name:<print name>`\n- You can remove a print from the list with `/print complete <print name>`", inline=False)
     embed.add_field(name='API Code', value=f'You can programatically interact with your stored prints via [your API code]({base_url + "/Saved/" + token}).\nNote that anyone can edit your stored prints using this API code.\nDo not share this with others.', inline=False)
     embed.add_field(name='Supported sites', value="Thingiverse, MyMiniFactory and Printables are supported by this bot. If you want more sites to be added, please see the contact info below.", inline=False)
     embed.add_field(name='Contact', value='If you have any questions, please contact [suchmememanyskill on Discord](https://discord.com/users/249186838592487425).\nThe source code of the bot can be found [on Github](https://github.com/suchmememanyskill/3d-print-queue-discord-bot).', inline=False)
+    embed.add_field(name='Stats', value=f"- {len(CHANNEL_MAPPINGS)} users have used this bot\n- {len(CACHE)} lists have been cached", inline=False)
 
     await interaction.followup.send(embed=embed, ephemeral=True)
 
 @bot.print_group.command(name='info', description='Embeds a 3d print url')
+async def print_info_command(interaction: discord.Interaction, url : str, show_in_channel : bool = False):
+    await print_info(interaction, url, show_in_channel)
+
 async def print_info(interaction: discord.Interaction, url : str, show_in_channel : bool = False):
     url = url.strip()
     await interaction.response.defer(ephemeral=not show_in_channel)
@@ -331,7 +331,7 @@ async def print_info(interaction: discord.Interaction, url : str, show_in_channe
         return
 
     embed = await uid_embed(uid)
-    view = OnlyDownloadButton(uid)
+    view = InteractButton(uid)
     await interaction.followup.send(embed=embed, view=view, ephemeral=not show_in_channel)
 
 @bot.event
